requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-HTTPAuth==4.8.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
Werkzeug==2.3.7
wsq==0.5
